Break apart zeroconf integration to prepare for WebSocket API
