+++ homeassistant/components/zeroconf/const.py
@@ -1,0 +1,5 @@
+"""Zeroconf constants."""
+
+DOMAIN = "zeroconf"
+
+ZEROCONF_TYPE = "_home-assistant._tcp.local."
